line-bot/requirements.txt
flask>=3.0.0,<4.0.0
line-bot-sdk>=3.5.0,<4.0.0
requests>=2.31.0,<3.0.0
